+++ src/registry.h
@@ -158,7 +158,7 @@ public:
   AntimonyEvent* GetCurrentEvent();
 
   //Algebraic rules
-  bool AddNewAlgebraicRuleToCurrent(int val, Formula* formula);
+  bool AddNewAlgebraicRuleToCurrent(double val, Formula* formula);
 
   Variable* GetCurrentSubmodel();
   void SetError(std::string error) {m_error = error;};
@@ -207,15 +207,15 @@ public:
   // CV terms
   bool ProcessGlobalCVTerm(const std::string* name, const std::string* qual, std::vector<std::string>* resources);
 
-  bool ProcessCreatorTerm(Annotated* a, const std::string* creator, const std::string* cterm, int resource);
+  bool ProcessCreatorTerm(Annotated* a, const std::string* creator, const std::string* cterm, double resource);
 
   bool ProcessCreatorTerm(Annotated* a, const std::string* creator, const std::string* cterm, std::vector<std::string>* resources);
 
-  bool ProcessGlobalCreatorTerm(const std::string* name, const std::string* creator, const std::string* cterm, int resource);
+  bool ProcessGlobalCreatorTerm(const std::string* name, const std::string* creator, const std::string* cterm, double resource);
 
   bool ProcessGlobalCreatorTerm(const std::string* name, const std::string* creator, const std::string* cterm, std::vector<std::string>* resources);
 
-  bool CheckCreatorString(const std::string& qualifier, int& creator_number);
+  bool CheckCreatorString(const std::string& qualifier, unsigned int& creator_number);
 
   //Keeping track of malloc'd stuff so we can free it ourselves if need be.
   std::vector<char*>    m_charstars;
